Adds goReleaser for release creation and binary builds
Adds [goReleaser](https://goreleaser.com/) support for creating binary
builds and releases in a standardized way.

Maintainers need to generate a Github token with the `repo` scope, and
save it in `~/.config/goreleaser/token`.  Then builds can be created by
running `make build`, and a release can be created from the most recent
tag by running `make release`.

Binaries will be built, a release created in Github, and the binaries
added to a *.tar.gz file with any README.md and LICENSE files.

An sha256sum.txt file is generated for each of the artifacts, and added
to the release.

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>