add --skip-version-check flag (#314)
* don't prompt to upgrade if running currentVersion-next

When compiling from the latest source code, osdctl would prompt the user
to upgrade to the latest version because it automatically appends
"-next" during compilation. Since the latest commit in the master branch
is newer than the latest tagged build, osdctl shouldn't prompt the user
to downgrade.

* remove dead code

* add an option to skip the latest-released-version check

The check was moved out of main() and into a PersistentPreRun hook on
the root command to take advantage of cobra argument parsing and keep
main() strictly for setting up and executing the command