Default to OCM login via environment
`osdctl` currently expects an OCM connection in order to run, but it
contains the packages necessary to build that connection automatically
based on the environment, as other OCM integrations do.

This commit adds support for automatically building the OCM connection
based on the presence of the `OCM_TOKEN` and `OCM_URL` environment
variables.

This will fall back to the existing connection usage if these environment
variables are not set.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>