osdctl cluster resize infra
This commit automates the machinepool dance that was being manually
followed to resize a classic OSD/ROSA cluster's infra nodes. It
automatically will select the next larger instance type up to
r5.8xlarge, but also supports supplying arbitrary instance types to
support edge cases, including reducing the size on infra nodes.

Notably, this does not automate creating/removing silences during this
process, nor the bookeeping tasks SRE's are tasked with doing after a
resize is complete, so the logs ask for confirmation before starting and
display a reminder for SRE's at the end to perform the remaining manual
steps.

Signed-off-by: Michael Shen <mshen@redhat.com>