Fixes PD retrieval bug; clean up long format output; other fixes
This PR fixes a bug in the pager duty alert retrieveal for the cluster
context command.

It also fixes a duplicate printing of the limited support reason where
support exceptions should be listed.

It also cleans up the log format output:

* Changes delimiters for more compact output
* standardizes line breaks
* removes "not polling for cloudtrail" message
* cleans up "external data" links
* replaces "Cluster Info" generic head with cluster name header, as
  short-form output uses

Finally, this PR addresses a handfull of spelling mistakes and renames
some variables to Golang convention (and helps with cSpell).

Cleans up the formatting for external links, standardizes printing

Signed-off-by: Chris Collins <collins.christopher@gmail.com>