Add default region for STS GetCallerIdentity in rotate-creds
This PR adds a region to the client used for the STS GetCallerIdentity and the GetAssumeRoleCredentials calles in the rotate-creds command.  This is required because the GetCallerIdentity call is expecting a region to be set, but there may not be one in the SRE's environment.  The region is unused for all of the calls the client is actually used for, and other clients in this command already hard-code `us-east-1` for their calls.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>