Fix the linting config and a selection of failures After reviewing our prow config I noticed that we are never actually running linting, which means this project has gotten really behind in linting failures.
I upgraded to `golangci-lint` v2, including the config file we use in `boilerplate`, and
configured it to ignore any failures prior to the latest commit for the project.

The reason I did that is I spent 2 hours fixing failures and there are still so many. This way
we can fix them in batches, while ensuring any new PRs start getting linting.

Once this merges I will open a PR on `openshift/release` to fix the lint target.