Adds helper info to `promotion` commit error message; prints commit message
Moves printing of the commit message above the attempt to commit, so that it is still printed even if an error occurs in the commit.  Adds a note that a failed commit may still succeed manually (for which the printed commit message is useful).

This is helpful in the case that, eg, you are using a toolbox, or ocm-contianer, and your GPG or SSH keys are not mounted.  The staged file can still be committed via a session with the required keys.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>