Adds provision shard ID and hive name to output
This adds the ability to get the provision shard ID for a given MC in
the case that you may want to pin a cluster install to a specific MC for
testing or development purposes.

Additionally this adds the hive that manages the MC to the output. This
will be helpful for finding which MCs are affected during soak periods
for MCC.